+++ src/globals.ts
@@ -3,3 +3,5 @@ export const API_V2_LABS_PREFIX = 'https://api.x.com/labs/2/';
 export const API_V1_1_PREFIX = 'https://api.x.com/1.1/';
 export const API_V1_1_UPLOAD_PREFIX = 'https://upload.x.com/1.1/';
 export const API_V1_1_STREAM_PREFIX = 'https://stream.x.com/1.1/';
+export const API_ADS_PREFIX = 'https://ads-api.x.com/12/';
+export const API_ADS_SANDBOX_PREFIX = 'https://ads-api-sandbox.twitter.com/12/';