+++ src/v2/client.v2.write.ts
@@ -20,9 +20,11 @@ import type {
   UserV2MuteResult,
   UserV2UnfollowResult,
   TweetV2BookmarkResult,
+  EUploadMimeType,
 } from '../types';
 import TwitterApiv2LabsReadWrite from '../v2-labs/client.v2.labs.write';
 import { CreateDMConversationParams, PostDMInConversationParams, PostDMInConversationResult } from '../types/v2/dm.v2.types';
+import { MediaV2MediaCategory, MediaV2MetadataCreateParams, MediaV2MetadataCreateResult, MediaV2UploadAppendParams, MediaV2UploadFinalizeParams, MediaV2UploadInitParams, MediaV2UploadResponse } from '../types/v2/media.v2.types';
 
 /**
  * Base Twitter v2 client with read/write rights.
@@ -119,6 +121,107 @@ export default class TwitterApiv2ReadWrite extends TwitterApiv2ReadOnly {
     }
 
     return this.post<TweetV2PostTweetResult>('tweets', payload);
+  }
+
+  /**
+   * Uploads media to Twitter using chunked upload.
+   * https://docs.x.com/x-api/media/media-upload
+   *
+   * @param media The media buffer to upload
+   * @param options Upload options including media type and category
+   * @param chunkSize Size of each chunk in bytes (default: 1MB)
+   * @returns The media ID of the uploaded media
+   */
+  public async uploadMedia(
+    media: Buffer,
+    options: { media_type: `${EUploadMimeType}` | EUploadMimeType; media_category?: MediaV2MediaCategory },
+    chunkSize: number = 1024 * 1024
+  ): Promise<string> {
+    let media_category = options.media_category;
+    // If no media category is provided, try to infer it from the media type
+    if (!options.media_category) {
+      if (options.media_type.includes('gif')) {
+        media_category = 'tweet_gif';
+       } else if (options.media_type.includes('image')) {
+        media_category = 'tweet_image';
+      } else if (options.media_type.includes('video')) {
+        media_category = 'tweet_video';
+      }
+    }
+
+    const initArguments: MediaV2UploadInitParams = {
+      command: 'INIT',
+      media_type: options.media_type,
+      total_bytes: media.length,
+      media_category,
+    };
+
+    const initResponse = await this.post<MediaV2UploadResponse>('media/upload', initArguments, { forceBodyMode: 'form-data' });
+    const mediaId = initResponse.data.id;
+
+    const chunksCount = Math.ceil(media.length / chunkSize);
+    const mediaArray = new Uint8Array(media);
+    for (let i = 0; i < chunksCount; i++) {
+      const start = i * chunkSize;
+      const end = Math.min(start + chunkSize, media.length);
+      const mediaChunk = mediaArray.slice(start, end);
+      const chunkedBuffer = Buffer.from(mediaChunk);
+
+      const appendArguments: MediaV2UploadAppendParams = {
+        command: 'APPEND',
+        media_id: mediaId,
+        segment_index: i,
+        media: chunkedBuffer,
+      };
+
+      await this.post('media/upload', appendArguments, { forceBodyMode: 'form-data' });
+    }
+
+    const finalizeArguments: MediaV2UploadFinalizeParams = {
+      command: 'FINALIZE',
+      media_id: mediaId,
+    };
+
+    const finalizeResponse = await this.post<MediaV2UploadResponse>('media/upload', finalizeArguments, { forceBodyMode: 'form-data' });
+    if (finalizeResponse.data.processing_info) {
+      await this.waitForMediaProcessing(mediaId);
+    }
+
+    return mediaId;
+  }
+
+  private async waitForMediaProcessing(mediaId: string): Promise<void> {
+    const response = await this.get<MediaV2UploadResponse>('media/upload', {
+      command: 'STATUS',
+      media_id: mediaId,
+    });
+
+    const info = response.data.processing_info;
+    if (!info) return;
+
+    switch (info.state) {
+      case 'succeeded':
+        return;
+      case 'failed':
+        throw new Error(`Media processing failed: ${info.error?.message}`);
+      case 'pending':
+      case 'in_progress': {
+        const waitTime = info?.check_after_secs;
+        if(waitTime && waitTime > 0) {
+          await new Promise(resolve => setTimeout(resolve, waitTime * 1000));
+          await this.waitForMediaProcessing(mediaId);
+        }
+      }
+    }
+  }
+
+  /**
+   * Creates the metadata for media to be uploaded.
+   * This feature is currently only supported for images and GIFs.
+   * https://docs.x.com/x-api/media/metadata-create
+   */
+  public createMediaMetadata(mediaId: string, metadata: Partial<MediaV2MetadataCreateParams>) {
+    return this.post<MediaV2MetadataCreateResult>('media/metadata', { id: mediaId, metadata });
   }
 
   /**